% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_stats.R
\name{bootstrap_maia}
\alias{bootstrap_maia}
\title{bootstrapping maia norm data}
\usage{
bootstrap_maia(n = 50, remove_diff = TRUE)
}
\arguments{
\item{n}{number of bootstrap samples}

\item{remove_diff}{logical. TRUE will remove cr_diff data (default)}
}
\value{
List
}
\description{
creates n bootstrap samples of the original norm_data
}
\details{
Returns list of linear models
}
\examples{
boots <- bootstrap_maia()
make linear model for each bootstrap sample
lm_boots <- lapply(boots, function(x) lm_loc(normdata = x))
# predict the location for each linear model.
# Using data of one test. But works also without testdata.
pred_boots <- lapply(
  lm_boots,
  function(x) predict_norm(testdata = testdat1, list_model = x)
)
# Creating the coordinates in testdat1 for the use in
# interpolate_norm
testdat_coord <- coord_cart(testdat1)

# now create the interpolation maps for each bootstrap sample
interpol_boots <- lapply(
  pred_boots, function(x) {
    interpolate_norm(pred_data = x, newgrid = testdat_coord)
    }
)
}
\seealso{
Other stat functions: 
\code{\link{mpstats_manual}()},
\code{\link{mpstats_single}()},
\code{\link{mpstats}()}
}
\author{
tjebo
}
\concept{stat functions}
