% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_stats.R
\name{calc_bebie}
\alias{calc_bebie}
\title{calculates bebie statistics}
\usage{
calc_bebie(testdata)
}
\arguments{
\item{testdata}{test for which field variation will be calculated}
}
\value{
List
}
\description{
linear regression model for each test location
of the grid underlying \link{norm_data}, with age as predictor
The output of this function is used for the prediction in
\link{predict_norm}
}
\details{
field variations should be created with \link{field_variations}
}
\examples{
field_var <- field_variation()
bebie_stats <- calc_bebie(testdat1, field_var)
}
\seealso{
Other prediction functions: 
\code{\link{coord_cart}()},
\code{\link{field_variation}()},
\code{\link{interpolate_norm}()},
\code{\link{lm_loc}()},
\code{\link{predict_norm}()}
}
\author{
tjebo
}
\concept{prediction functions}
