% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_globInd.R
\name{calc_globInd}
\alias{calc_globInd}
\title{calc_globInd}
\usage{
calc_globInd(df, test_val, norm_val, var_norm)
}
\arguments{
\item{df}{Dataframe in which the vectors are found}

\item{test_val}{observed sensitivities (vector)}

\item{norm_val}{normal sensitivities for the same points}

\item{var_norm}{variance of normal values for the same points}
}
\value{
named vector with mean deviation (MDev) and pattern standard deviation (PSD)
}
\description{
calculates global indices (mean deviation and pattern standard deviation) from given vectors
}
\author{
tjebo
}
