% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maia_tgz.R
\name{clean_mp_data}
\alias{clean_mp_data}
\title{clean_mp_data}
\usage{
clean_mp_data(x, timeclass)
}
\arguments{
\item{x}{data frame}

\item{timeclass}{either 'date' (default) or 'datetime' (POSIXct)}
}
\value{
Data frame
}
\description{
cleaning up extracted data frame with mp test data and PII
}
