% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{compare}
\alias{compare}
\title{compare}
\usage{
compare(testdata)
}
\arguments{
\item{testdata}{\strong{required}. Data frame with maia test results. See details}
}
\value{
data frames with original and predicted normal values.
}
\description{
Interpolates normal values and compares with test values.
This is basically a wrapper around \link{predict_norm}, \link{coord_cart} and
\link{interpolate_norm}
}
\details{
\strong{testdata}: This should be ideally output from \link{read_maia}.
The data frame must contain following columns:
\strong{patID, eye, testID, testtype, eccent, angle, value, stimID}
}
\author{
tjebo
}
