% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{coord_cart}
\alias{coord_cart}
\title{coord_cart}
\usage{
coord_cart(testdata)
}
\arguments{
\item{testdata}{\strong{required}. Data frame, output from \link{read_maia}.}
}
\value{
Data frames
}
\description{
Adding cartesian coordinates to test results
}
\examples{
preddat <- predict_norm(testdat1)
testdat_coord <- coord_cart(testdat1)
interpol_dat <- interpolate_norm(preddat, newgrid = testdat_coord)
}
\seealso{
Other prediction functions: 
\code{\link{interpolate_norm}()},
\code{\link{predict_norm}()}
}
\concept{prediction functions}
