% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maia_tgz.R
\name{extract_pii}
\alias{extract_pii}
\title{extracts PII}
\usage{
extract_pii(x)
}
\arguments{
\item{x}{jsonfile}
}
\value{
Data frame
}
\description{
Extracts personally identifiable information from MAIA raw data.
}
\details{
Internal function to extract name, dob and and sex from MAIA observers.
This data is stored in json files attached to the microperimetry files.
It is generally not recommended to use this function explicitely in order to keep personal identifiable data as safe as possible.
}
\author{
Tjebo
}
\keyword{internal}
