% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{field_variation}
\alias{field_variation}
\title{variation of the visual field}
\usage{
field_variation()
}
\value{
data frame
}
\description{
generating field variations based on simple kriging
of normal fields, without the step of predicting location with
linear models first.
}
\examples{
field_var <- field_variation()
bebie_stats <- calc_bebie(testdat1, field_var)
}
\seealso{
Other prediction functions: 
\code{\link{calc_bebie}()},
\code{\link{coord_cart}()},
\code{\link{interpolate_norm}()},
\code{\link{lm_loc}()},
\code{\link{predict_norm}()}
}
\author{
tjebo
}
\concept{prediction functions}
