% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{get_names
Extract patient identifiable information from MAIA raw data}
\usage{
get_names(folder = getwd())
}
\arguments{
\item{folder}{source folder which is searched for .tgz files. Default: workdirectory}
}
\value{
Data frame
}
\description{
Internal function to extract name, dob and and sex from MAIA observers.
This data is stored in json files attached to the microperimetry files.
It is generally not recommended to use this function explicitely in order to keep personal identifiable data as safe as possible.
}
\author{
Tjebo
}
