% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{interpolate_norm}
\alias{interpolate_norm}
\title{interpolate_norm}
\usage{
interpolate_norm(
  age,
  sex,
  lens = "natural",
  interval = "confidence",
  as_df = FALSE,
  grid_density = 0.2,
  graph = FALSE
)
}
\arguments{
\item{age}{required age of individual to be predicted}

\item{sex}{required sex of individual to be predicted}

\item{lens}{lens status. must be either 'natural' (default) or 'pseudo'}

\item{interval}{used in [stats::predict.lm()]}

\item{as_df}{FALSE (default): list of data frames will be returnedif TRUE, value of data frame will be returned.}

\item{grid_density}{resolution of interpolated grid in degress.}

\item{graph}{if TRUE, the interpolation will be plotted for all test types}
}
\value{
List of data frames with predicted results for grid for all test types
and plot if graph = TRUE
# Tests were performed
 plot(gstat::variogramLine(fit_fit, 10), type='l')
 points(vario[,2:3], pch=20, col='red')
LOOCV of data points from interpol_dat data
x <- gstat::krige.cv(fit ~ 1, interpol_dat, interpol_dat, model = fit_fit, nfold = nrow(interpol_dat))
RMSE(x$var1.pred,x$observed)
vgm.fit
plot(vgm, vgm.fit)
sp::spplot(mes_ok_mean ["var1.pred"])
}
\description{
Interpolates grid based on local predictions of normal data.
}
\author{
tjebo and max
}
