% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{lm_loc}
\alias{lm_loc}
\title{linear model for each location}
\usage{
lm_loc(testdata = NULL, normdata = NULL)
}
\arguments{
\item{testdata}{if given, the lm will only be created for testtype
of the testdata}

\item{normdata}{data used for prediction, defaults to norm_data.}
}
\value{
List
}
\description{
linear regression model for each test location
of the grid underlying \link{norm_data}, with age as predictor
The output of this function is used for the prediction in
\link{predict_norm}
}
\details{
Returns list of linear models for all test types
at each location of the grid underlying \link{norm_data}
}
\examples{
lm_dat <- lm_loc(testdata = testdat1)
preddat <- predict_norm(testdat1, list_model = lm_dat)
testdat_coord <- coord_cart(testdat1)
interpol_dat <- interpolate_norm(preddat, newgrid = testdat_coord)
}
\seealso{
Other prediction functions: 
\code{\link{calc_bebie}()},
\code{\link{coord_cart}()},
\code{\link{field_variation}()},
\code{\link{interpolate_norm}()},
\code{\link{predict_norm}()}
}
\author{
tjebo
}
\concept{prediction functions}
