% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maia_tgz.R
\docType{data}
\name{lu_test_type}
\alias{lu_test_type}
\title{clean_mp_data}
\format{
An object of class \code{character} of length 4.
}
\usage{
lu_test_type
}
\arguments{
\item{x}{data frame}

\item{timeclass}{either 'date' (default) or 'datetime' (POSIXct)}

\item{incomplete}{set TRUE, if you want to see incomplete exams too}
}
\value{
Data frame
}
\description{
cleaning up extracted data frame with mp test data and PII
}
\keyword{datasets}
