% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{maia_palette}
\alias{maia_palette}
\title{maia_palette}
\format{
An object of class \code{character} of length 4.
}
\usage{
maia_palette
}
\description{
color palette for maia colors
}
\keyword{datasets}
