% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maiaR.R
\docType{package}
\name{maiaR}
\alias{maiaR}
\title{maiaR}
\description{
For working with MAIA raw data
}
\examples{
# Guide to plotting normal sensitivity data
# Includes linear regression lines
ggplot(norm_data, aes(x = Age, y = value)) +
  geom_point(aes(x = Age, y = value, shape = lens), alpha = 0.05, show.legend = FALSE) +
  facet_wrap(~ testtype) +
  ylab("Sensitivity [dB]") +
  xlab("Age [years]") +
  geom_smooth(aes(color = testtype, linetype = lens), method = 'lm', formula = y ~ x) +
  scale_color_manual(name = 'Regression lines', values = maia_palette)
}
\author{
Tjebo Heeren and Maximilian Pfau
}
