% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maia_tgz.R
\name{make_df_each_xml}
\alias{make_df_each_xml}
\title{extract xml data}
\usage{
make_df_each_xml(x, incomplete)
}
\arguments{
\item{x}{name of XML file}

\item{incomplete}{set TRUE, if you want to see incomplete exams too}
}
\value{
Data frame
}
\description{
core function to create data frame from xml attributes
}
\keyword{internal}
