% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{MD_PSD}
\alias{MD_PSD}
\title{MD_PSD}
\usage{
MD_PSD(data)
}
\value{
named vector with mean deviation (MDev) and pattern standard deviation (PSD)
}
\description{
calculates estimated mean deviation (MD). This is estimated, because there is no 'real mean' of normal values for each location.
The mean and variance are taken from the local linear regression taken from the normal data, including age, sex and lens status as covariates.
Variance is estimated with the \link[stats:predict.lm]{stats::predict.lm} function, assuming homoscedasticity in a non-weighted linear regression with equal variance or residuals.
The estimated variance is (currently) equal to the prediction interval
}
\author{
tjebo
}
