% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{MD_PSD_man}
\alias{MD_PSD_man}
\title{MD_PSD_man}
\usage{
MD_PSD_man(df, test_val, norm_val, var_norm)
}
\arguments{
\item{df}{Dataframe in which the vectors are found}

\item{test_val}{observed sensitivities (either column name or vector)}

\item{norm_val}{normal sensitivities for the same points (either column name or vector)}

\item{var_norm}{variance of normal values for the same points (either column name or vector)}
}
\value{
named vector with mean deviation (MDev) and pattern standard deviation (PSD)
}
\description{
calculates global indices mean deviation (MD) and pattern standard deviation (PSD) manually from data or vectors
}
\author{
tjebo
}
