% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\docType{data}
\name{MeanDev}
\alias{MeanDev}
\title{MeanDev}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
MeanDev
}
\value{
named vector with mean deviation (MDev) and pattern standard deviation (PSD)
}
\description{
calculates estimated mean deviation (MD)
Variance is estimated with the \link[stats:predict.lm]{stats::predict.lm} function, assuming homoscedasticity
(non-weighted linear regression with equal variance or residuals) residual.scale ^ 2.
}
\author{
tjebo
}
\keyword{datasets}
