% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_stats.R
\name{meanSens}
\alias{meanSens}
\title{Microperimetry basic statistics.}
\usage{
meanSens(data, global = FALSE, na.rm = TRUE, digits)
}
\arguments{
\item{data}{microperimetry test data, ideally import with read_maia_tgz}

\item{global}{TRUE (MS and SD of all tests will be calculated) or
FALSE (default), where MS and SD will be provided for each test separately}

\item{na.rm}{TRUE (default): NA will be ignored. False: will return NA}

\item{digits}{round to how many digits.}
}
\value{
if global = TRUE: vector of MS and SD, else dataframe (or: tibble)
}
\description{
Calculates mean sensitivity (MS) and standard deviation (SD) from
all tests within a data frame containing microperimetry test data.
}
