% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_stats.R
\name{mpstats_manual}
\alias{mpstats_manual}
\title{"Manual" calculation of microperimetry statistics}
\usage{
mpstats_manual(test_val, norm_val, var_norm, data = NULL)
}
\arguments{
\item{test_val}{observed sensitivities. Can be column name (quoted or unquoted) or numeric vector}

\item{norm_val}{mean of normal sensitivities of same length as test_val. Can be column name (quoted or unquoted) or numeric vector}

\item{var_norm}{variance of normal values of same length as test_val. Can be column name (quoted or unquoted) or numeric vector}

\item{data}{\strong{optional} Dataframe with vectors for analysis}
}
\value{
named vector with mean deviation (mean_dev) and pattern standard deviation (psd)
}
\description{
Calculates mean deviation (MD) and pattern standard deviation (psd)
manually from data or vectors.
}
\seealso{
Other stat functions: 
\code{\link{mpstats_single}()},
\code{\link{mpstats}()}
}
\author{
tjebo
}
\concept{stat functions}
