% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_stats.R
\name{mpstats_single}
\alias{mpstats_single}
\title{Microperimetry statistics.}
\usage{
mpstats_single(interpol_dat, digits = 2)
}
\arguments{
\item{interpol_dat}{interpolated results from test data.}

\item{digits}{Digits to be rounded to}
}
\value{
vector
}
\description{
Calculates microperimetry statistics from a single test.
Includes mean sensitivity (mean_sens), mean deviation (mean_dev),
pattern standard deviation (psd). See also details.
Works in conjunction with prediction functions. (see example)
}
\details{
\strong{mean_dev} and \strong{psd} are estimated.
This is, because there is no 'real mean' and, more importantly,
no "real variance" of normal values for each given location.
Mean and variance are estimated based on kringing (spatial interpolation)
of \emph{predicted (!)} values. The predicted values are given by simple linear
regression for each location, based on the \link{norm_data}.
It includes age, sex and lens status as covariates.
The variance used for estimation is equal to the \emph{prediction intervals}
from those models, assuming homoscedasticity in a non-weighted model
}
\examples{
preddat <- predict_norm(testdat1)
testdat_coord <- coord_cart(testdat1)
interpol_dat <- interpolate_norm(preddat, newgrid = testdat_coord)
mpstats_single(interpol_dat)
}
\seealso{
\link[stats:predict.lm]{stats::predict.lm}

Other stat functions: 
\code{\link{mpstats_manual}()},
\code{\link{mpstats}()}
}
\author{
tjebo
}
\concept{stat functions}
