% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{norm_data}
\alias{norm_data}
\title{Norm data}
\format{
A data frame with \Sexpr{nrow(maiaR::norm_data)} rows and NN variables:
}
\source{
University Eye Hospital Bonn. Collected by Simone Tzaridis and Maximilian Pfau
}
\usage{
data(norm_data)
}
\description{
Data obtained from healthy controls. A broad range of ages were used.
It contains data of \Sexpr{unique(interaction(maiaR::norm_data$patID, maiaR::norm_data$eye))} eyes of \Sexpr{unique(maiaR::norm_data$patID)} observers.
}
\details{
DOCUMENTATION TBC

\itemize{
   \item patID
   }
}
