% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_bebie}
\alias{plot_bebie}
\title{Plotting bebie curves}
\usage{
plot_bebie(data, ...)
}
\arguments{
\item{data}{Data for which bebie curve will be calculated. Can be either
the original test data or the output from \link{calc_bebie}}

\item{...}{further arguments which will be passed to geom_line for the
test data}
}
\value{
ggplot object with side effect of printing to console
}
\description{
Plotting bebie curve. For details of calculation of bebie stats,
see \link{field_variation} and \link{calc_bebie}
}
\examples{
field_var <- field_variation(testdat1)
bebie_stats <- calc_bebie(testdat1, field_var)
plot_bebie(bebie_stats)
}
\seealso{
\link[ggplot2:geom_line]{ggplot2::geom_line}

Other microperimetry plotting functions: 
\code{\link{plot_bland}()},
\code{\link{plot_mpstats}()}
}
\author{
tjebo
}
\concept{microperimetry plotting functions}
