% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_bland}
\alias{plot_bland}
\title{Plotting bland altman}
\usage{
plot_bland(testdata = NULL, ...)
}
\arguments{
\item{testdata}{needs to be data frame with two columns for each test.
If not given, the data from sMAIA test retest will be shown}

\item{...}{further arguments which will be passed to geom_point}
}
\value{
ggplot object with side effect of printing to console
}
\description{
Plotting bland altman plot of test retest of sMAIA device
Data from norm data
}
\seealso{
\link{geom_point}
}
\author{
tjebo
}
