% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mpstats}
\alias{plot_mpstats}
\title{plot_mpstats}
\usage{
plot_mpstats(testdata, quantiles = 0.1)
}
\arguments{
\item{testdata}{test data. Do not try to pass a list.
Built to pass the output from \link{read_maia}}

\item{quantiles}{Quantiles for normal values to be displayed.}
}
\value{
List of plots which are bound together with the patchwork package.
}
\description{
plotting relative percentile of estimated Mean deviation for tested observers
}
\examples{
# testdata <- read_maia(folder = file.path(getwd(), "data-raw"))
# plot_mpstats(testdata)
}
\seealso{
Other microperimetry plotting functions: 
\code{\link{plot_bebie}()},
\code{\link{plot_bland}()}
}
\author{
tjebo
}
\concept{microperimetry plotting functions}
