% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_PSD}
\alias{plot_PSD}
\title{plot_PSD}
\usage{
plot_PSD(testdata, quantiles = 0.1)
}
\arguments{
\item{testdata}{test data. Do not try to pass a list. Built to pass the output from \link{read_maia} or, even better, from \link{compare_norm}
If data is organised in list of data frames, bind those first.}

\item{quantiles}{Quantiles for normal values to be displayed.}
}
\value{
List of plots which are bound together with the patchwork package.
}
\description{
plotting relative percentile of estimated Mean deviation for tested observers
}
\examples{
testdata <- read_maia()
comparedat <- compare_norm(testdata)
plot_PSD(comparedat)
}
\author{
tjebo
}
