% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{pred_location}
\alias{pred_location}
\title{pred_location}
\usage{
pred_location(age, sex, lens = "natural", interval = "confidence")
}
\arguments{
\item{age}{required age of individual to be predicted}

\item{sex}{required sex of individual to be predicted}

\item{lens}{lens status. must be either 'natural' (default) or 'pseudo'}

\item{interval}{used in [stats::predict.lm()]}
}
\value{
Data frame with predicted sensitivity values for all test types at each location
}
\description{
location based prediction of retinal sensitivity using a linear model for each location with age, sex and lens status as covariates
Note this function is used under the hood for the prediction of interpolated values.
You can use it for prediction of retinal sensitivity at one of the test locations which underly the normal data, but it would probably make more sense to create a mixed effect model, taking into account the within-subject-variance
E.g. lme4::lmer(value ~ age + sex + lens + (1|patID))
}
\author{
tjebo
}
