% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maia.R
\name{read_maia}
\alias{read_maia}
\title{read_maia
Extract MAIA raw data}
\usage{
read_maia(folder = getwd(), incomplete = FALSE, timeclass = "datetime")
}
\arguments{
\item{folder}{source folder which is searched for .tgz files. Default: workdirectory}

\item{incomplete}{set TRUE, if you want to see incomplete exams too}

\item{timeclass}{either 'date' (date class) or 'datetime' (POSIXct class), for date and time of test}
}
\value{
Data frame
}
\description{
The function will extract the anonymised data from the raw .tgz output of your MAIA.
}
\author{
Tjebo
}
