% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_fct.R
\name{RMSE}
\alias{RMSE}
\title{RMSE}
\usage{
RMSE(observed, predicted)
}
\arguments{
\item{observed}{vector of observed values}

\item{predicted}{vector of predicted values}
}
\value{
numeric vector
}
\description{
calculates root mean square error (RMSE) for model testing
}
