% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maia_tgz.R
\name{unzip_tgz}
\alias{unzip_tgz}
\title{unzips tgz files}
\usage{
unzip_tgz(tgz_file, tmpdir)
}
\arguments{
\item{tgz_file}{file to extract}

\item{tmpdir}{temporary directory}
}
\description{
unzips tgz files
}
