% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/using.R
\name{using}
\alias{using}
\title{#' using}
\source{
https://stackoverflow.com/a/44660688/7941188
}
\usage{
using(...)
}
\description{
internal function for test of package presence.
}
\author{
SO user Matthew https://stackoverflow.com/users/4125693/matthew
}
