% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{vec_normpositions}
\alias{vec_normpositions}
\title{vec_normpositions}
\format{
An object of class \code{character} of length 71.
}
\usage{
vec_normpositions
}
\description{
vector with positions in order to create level order
}
\keyword{datasets}
