% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_carts.R
\name{add_carts}
\alias{add_carts}
\title{add_carts}
\usage{
add_carts(data, digits = 2)
}
\arguments{
\item{data}{Data frame, ideally output from \link{read_maia_tgz}.}

\item{digits}{how precise to be about eccentricity and angles}
}
\value{
Data frames
}
\description{
Adding cartesian coordinates to test results
}
\details{
if not using output from read_maia_tgz, the data frame needs to contain a
column called "angle" (angle given in radians) and "eccent" (eccentricity given in any unit)
}
\examples{
lm_dat <- lm_loc(testdata = testdat1)
preddat <- predict_norm(testdat1, list_model = lm_dat)
testdat_coord <- coord_cart(testdat1)
interpol_dat <- interpolate_norm(preddat, newgrid = testdat_coord)
}
\concept{prediction functions}
