% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{compare_norm}
\alias{compare_norm}
\title{compare_norm}
\usage{
compare_norm(
  test_data,
  age = NULL,
  sex = NULL,
  lens = NULL,
  interval = "confidence"
)
}
\arguments{
\item{test_data}{\strong{required}. data frame with eccentricity and degree and corresponding test value}

\item{age}{age of individual to be predicted. passed to \link{pred_location}}

\item{sex}{sex of individual to be predicted. passed to \link{pred_location}}

\item{lens}{lens status. must be either 'natural' (default) or 'pseudo'. passed to \link{pred_location}}

\item{interval}{used in \link[stats:predict.lm]{stats::predict.lm}}
}
\value{
Data frame with original and predicted normal values.
}
\description{
Interpolates grid based on local predictions of normal data.
For test points that correspond with basic normal grid, the values from \link{pred_location} will be used.
For test points outside the normal grid location, values from \link{interpolate_norm} will be used
}
\author{
tjebo
}
