% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{compare_norm}
\alias{compare_norm}
\title{compare_norm}
\usage{
compare_norm(testresults, interval = "predict")
}
\arguments{
\item{interval}{used in \link[stats:predict.lm]{stats::predict.lm}}

\item{data}{\strong{required}. data frame with sensitivity values. See details.
Data argument should be data frame only! List of data frames could result in unexpected output.
The data frame is ideally taken directly from \link{read_maia} output, ideally joined with information on lens status. (not contained in read_maia() output!)
It should at least contain following columns: \strong{patID, eye, testID, testtype, eccent, angle, value, stimID}
It should ideally also contain \strong{age, sex and lens} - If not, defaults of age, sex and lens of \link{pred_location} will be used}
}
\value{
list of data frames (for each testID) with original and predicted normal values.
}
\description{
Interpolates grid based on local predictions of normal data.
For test points that correspond with basic normal grid, the values from \link{pred_location} will be used.
For test points outside the normal grid location, values from \link{interpolate_norm} will be used
}
\author{
tjebo
}
