% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{CoR_maia}
\alias{CoR_maia}
\title{Coefficient of repeatability}
\usage{
CoR_maia(graph = TRUE)
}
\arguments{
\item{graph}{if TRUE, bland altman plot will be printed (with error lines at +/- 1.96*sd)}
}
\value{
Data frame with coefficients of repeatability for all test types
and plot if graph = TRUE
}
\description{
Coefficient of repeatability of norm data
Calculated with formula CoR = 1.96*sqrt(2)* sqrt(within-subject-variance)
}
\author{
tjebo
}
