% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age.R
\name{get_age}
\alias{get_age}
\title{get the age (time difference in years)}
\usage{
get_age(from_date, to_date = lubridate::now(), dec = FALSE)
}
\arguments{
\item{from_date}{beginning of time span}

\item{to_date}{end (default: now!)}

\item{dec}{decide if you want to see decimals (default: FALSE)}
}
\value{
Data frame
}
\description{
The function will extract the data from the raw .tgz output of your MAIA.
}
\author{
Stackoverflow :: Moody Mudskipper
}
