% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMAIA.R
\name{getMaia}
\alias{getMaia}
\title{Extract MAIA raw data}
\usage{
getMaia(folder = getwd(), show_incomplete = FALSE,
  show_examDate = FALSE)
}
\arguments{
\item{folder}{source folder which is searched for .tgz files. Default: workdirectory}

\item{show_incomplete}{set TRUE, if you want to see incomplete exams too}

\item{show_examDate}{default = FALSE}
}
\value{
Data frame
}
\description{
The function will extract the data from the raw .tgz output of your MAIA.
}
\author{
Tjebo
}
