% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{interpolate_norm}
\alias{interpolate_norm}
\title{interpolate_norm}
\usage{
interpolate_norm(pred_data, newgrid = NULL, grid_density = 0.2)
}
\arguments{
\item{pred_data}{predicted norm data from \link{predict_norm}}

\item{newgrid}{custom grid to be interpolated on. If not specified, regular grid will be interpolated with resolution given in grid_density}

\item{grid_density}{resolution of interpolated grid in degress.}
}
\value{
List of data frames with predicted results for grid for all test types and plot if graph = TRUE
}
\description{
Interpolates grid based on local predictions of normal data.
}
\examples{
# Tests were performed with leave-one out cross validation
#LOOCV of data points from interpol_dat data
#x <- gstat::krige.cv(fit ~ 1, interpol_dat, interpol_dat,
#       model = fit_fit, nfold = nrow(interpol_dat))
#RMSE(x$var1.pred,x$observed)

lm_dat <- lm_loc(testdata = testdat1)
preddat <- predict_norm(testdat1, list_model = lm_dat)
testdat_coord <- coord_cart(testdat1)
interpol_dat <- interpolate_norm(preddat, newgrid = testdat_coord)
}
\seealso{
Other prediction functions: 
\code{\link{calc_bebie}()},
\code{\link{coord_cart}()},
\code{\link{field_variation}()},
\code{\link{lm_loc}()},
\code{\link{predict_norm}()}
}
\author{
tjebo and max
}
\concept{prediction functions}
