% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{MD_PSD_man}
\alias{MD_PSD_man}
\title{MD_PSD_man}
\usage{
MD_PSD_man(test_val, norm_val, var_norm, data = NULL)
}
\arguments{
\item{test_val}{observed sensitivities. Can be column name (quoted or unquoted) or numeric vector}

\item{norm_val}{mean of normal sensitivities of same length as test_val. Can be column name (quoted or unquoted) or numeric vector}

\item{var_norm}{variance of normal values of same length as test_val. Can be column name (quoted or unquoted) or numeric vector}

\item{data}{\strong{optional} Dataframe with vectors for analysis}
}
\value{
named vector with mean deviation (MDev) and pattern standard deviation (PSD)
}
\description{
calculates global indices mean deviation (MD) and pattern standard deviation (PSD) manually from data or vectors
}
\author{
tjebo
}
