% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microperimetR.R
\docType{package}
\name{microperimetR}
\alias{microperimetR}
\title{microperimetR}
\description{
For working with MAIA raw data
}
\section{plotting functions (TBC)}{
NA
}

\section{statistics and summary functions (TBC)}{
NA
}

\examples{
# Guide to plotting normal sensitivity data
# Includes linear regression lines
ggplot(norm_data, aes(x = Age, y = value)) +
  geom_point(aes(x = Age, y = value, shape = lens), alpha = 0.05, show.legend = FALSE) +
  facet_wrap(~ testtype) +
  ylab("Sensitivity [dB]") +
  xlab("Age [years]") +
  geom_smooth(aes(color = testtype, linetype = lens), method = 'lm', formula = y ~ x) +
  scale_color_manual(name = 'Regression lines', values = maia_palette)
}
\author{
Tjebo Heeren and Maximilian Pfau and ... TBC!
}
