% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{norm_data}
\alias{norm_data}
\title{Norm data}
\format{
A data frame with \Sexpr{nrow(maiaR::norm_data)} rows and NN variables:
}
\source{
University Eye Hospital Bonn. Collected by Maximilian Pfau
}
\usage{
data(norm_data)
}
\description{
Data obtained from healthy controls. A broad range of ages were used.
It contains data of n eyes of n observers.
}
\details{
\itemize{
   \item patID
   \item age
   \item baseID
   \item examID
   \item examType
   \item eye
   \item avg_rctn
   \item stimID
   \item fi
   \item ray
   \item a_deg
}
}
