% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_MD}
\alias{plot_MD}
\title{plot_MD}
\usage{
plot_MD(testdata)
}
\arguments{
\item{testdata}{test data. Do not try to pass a list. Ideally, pass the output from \link{read_maia}
If data is organised in list of data frames, bind those first.}
}
\value{
List of plots which are bound together with the patchwork package.
}
\description{
plotting relative percentile of estimated Mean deviation for tested observers
}
\examples{
testdata <- read_maia()
plot_MD(testdata)
}
\author{
tjebo
}
