% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_norm}
\alias{predict_norm}
\title{predict_norm}
\usage{
predict_norm(
  testdata = NULL,
  age = NULL,
  sex = NULL,
  lens = NULL,
  interval = "prediction"
)
}
\arguments{
\item{testdata}{if passed, the testtype of the test will be extracted}

\item{age}{age of individual to be predicted. defaults to 50.}

\item{sex}{sex of individual to be predicted. defaults to 'm'.}

\item{lens}{lens status. must be either 'natural' (default) or 'pseudo'.
Defaults to 'natural'}

\item{interval}{used in \link[stats]{predict.lm}}
}
\value{
Data frame with predicted sensitivity values for all test types
at each location of the grid underlying \link{norm_data}
}
\description{
location based prediction of retinal sensitivity
using a linear model for each location of the grid underlying \link{norm_data}
with age, sex and lens status
as covariates. The output of this function is used for the interpolation
with \link{interpolate_norm}
}
\examples{
preddat <- predict_norm(testdat1)
testdat_coord <- coord_cart(testdat1)
interpol_dat <- interpolate_norm(preddat, newgrid = testdat_coord)
}
\seealso{
Other prediction functions: 
\code{\link{coord_cart}()},
\code{\link{interpolate_norm}()}
}
\author{
tjebo
}
\concept{prediction functions}
