% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maia_tgz.R
\name{read_maia_tgz}
\alias{read_maia_tgz}
\title{Read MAIA files}
\usage{
read_maia_tgz(folder = getwd(), incomplete = FALSE, timeclass = "date")
}
\arguments{
\item{folder}{source folder which is searched for .tgz files. Default: workdirectory}

\item{incomplete}{set TRUE, if you want to see incomplete exams too}

\item{timeclass}{either 'date' (Date class, default) or 'datetime' (POSIXct), how to provide microperimetry test time}
}
\value{
Data frame
}
\description{
The function will extract the data from the raw .tgz patient backup files of your MAIA device.
}
\examples{
# read_maia_tgz(folder = file.path(getwd(), "data-raw"))
}
\author{
Tjebo
}
