% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maia_txt.R
\name{read_maia_txt}
\alias{read_maia_txt}
\title{Read MAIA data in .txt files}
\usage{
read_maia_txt(dir = getwd())
}
\arguments{
\item{dir}{source folder. Default: working directory (check with
getwd())}

\item{incomplete}{set TRUE, if you want to see incomplete exams too}

\item{timeclass}{either 'date' (date class) or 'datetime' (POSIXct class), for date and time of test}
}
\value{
Data frame
}
\description{
Extraction of data from exported .txt files
}
\examples{
# read_maia_txt()
}
\author{
Tjebo
}
